[Feat/DetailPage] 게시글 신고 코드 분리, 찜/신고 코드 정리
